+++ ibrl/environments/base_newcomb_like.py
@@ -30,7 +30,7 @@ class BaseNewcombLikeEnvironment(BaseEnvironment):
     def interact(self, action : int) -> float:
         return self.rewards[action]
 
-    def get_optimal_reward(self) -> int:
+    def get_optimal_reward(self) -> float:
         # Compute the optimal reward, based on the full reward table
         # The reward is a quadratic function of the probability of taking action 0.
         # Thus, there are three policies that could potentially be optimal
